int instead of size_t

+++ src/tests/c_api/test_join_sail.cpp
@@ -266,7 +266,7 @@ int main(int argc, char** argv) {
 
         // Send header vals  to P2
         for (size_t i = 0; i < COLS1; i++){
-            size_t vecSize = js1_header_toSend.size();
+            int vecSize = js1_header_toSend.size();
             MPI_Send(vecSize, 1, MPI_LONG_LONG, 1, HEADER_TAG, MPI_COMM_WORLD);
             MPI_Send(js1_header_toSend[i].data(), vecSize, MPI_LONG_LONG, 1, HEADER_TAG, MPI_COMM_WORLD);
         }
@@ -274,7 +274,7 @@ int main(int argc, char** argv) {
         // Receive P2's header and convert to string
         std::vector<std::string> js2_header;
         for (size_t i = 0; i < COLS2; i++){
-            size_t vecSize;
+            int vecSize;
             MPI_Recv(vecSize, 1, MPI_LONG_LONG, 1, HEADER_TAG, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
             std::vector<unsigned long long> curr_header(vecSize);
             MPI_Recv(curr_header.data(), vecSize, MPI_LONG_LONG, 1, HEADER_TAG, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
@@ -460,7 +460,7 @@ int main(int argc, char** argv) {
         // Receive P1's header and convert to string
         std::vector<std::string> js1_header;
         for (size_t i = 0; i < COLS1; i++){
-            size_t vecSize;
+            int vecSize;
             MPI_Recv(vecSize, 1, MPI_LONG_LONG, 0, HEADER_TAG, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
             std::vector<unsigned long long> curr_header(vecSize);
             MPI_Recv(curr_header.data(), vecSize, MPI_LONG_LONG, 0, HEADER_TAG, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
@@ -469,7 +469,7 @@ int main(int argc, char** argv) {
 
         // Send header vals to P1
         for (size_t i = 0; i < COLS2; i++){
-            size_t vecSize = js2_header_toSend.size();
+            int vecSize = js2_header_toSend.size();
             MPI_Send(vecSize, 1, MPI_LONG_LONG, 0, HEADER_TAG, MPI_COMM_WORLD);
             MPI_Send(js2_header_toSend[i].data(), vecSize, MPI_LONG_LONG, 0, HEADER_TAG, MPI_COMM_WORLD);
         }
